G2-55: Fix typos, add more control flow, and tweak momentarily for testing

+++ ejabberd_build_centos.sh
@@ -81,15 +81,15 @@ sudo iptables -I INPUT -p tcp --dport 4369 -m state --state NEW,ESTABLISHED -j A
 iptables-save
 
 # Clone and build ejabberd from source
-git clone https://github.com/processone/ejabberd.git /tmp/ejabberd
-cd /tmp/ejabberd
+#####git clone https://github.com/processone/ejabberd.git /tmp/ejabberd
+#####cd /tmp/ejabberd
 # Remove traces from a previous installation that might pose conflicts
-sudo make clean 
-./autogen.sh
+#####sudo make clean 
+#####./autogen.sh
 # Configure Ejabberd to use PostgreSQL 
-./configure --enable-pgsql
-make
-sudo make install 
+#####./configure --enable-pgsql
+#####make
+#####sudo make install 
 
 # Purely for testing in isolation, create postgres database
 cd $current_dir
@@ -212,7 +212,9 @@ sudo chown -R ejabberd:ejabberd /var/lock/ejabberdctl
 sudo chown ejabberd:ejabberd /sbin/ejabberdctl
 
 # Copy the erlang cookie into Ejabberd's home directory
-sudo cp ~/erlang.cookie /home/ejabberd/.erlang.cookie
+if [ -f ~/.erlang.cookie ]; then 
+	sudo cp ~/erlang.cookie /home/ejabberd/.erlang.cookie
+fi 
 sudo chown ejabberd:ejabberd /home/ejabberd/.erlang.cookie
 sudo chmod 400 /home/ejabberd/.erlang.cookie
